updated assignment 1 mistakes thru pylint
