+++ lessons/lesson01/assignment/inventory_management/electric_appliances_class.py
