+++ lessons/lesson01/assignment/inventory_management/furniture_class.py
