+++ lessons/lesson01/assignment/inventory_management/inventory_class.py
